requirements.txt
streamlit==1.31.0
tensorflow==2.14.0
numpy==1.24.3
Pillow==10.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
bcrypt==4.1.2
pandas==2.1.4
plotly==5.18.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
streamlit-option-menu==0.3.12
streamlit-aggrid==0.3.4 